drivers: refactor: move timer from platform to drivers
+ Indents for 'CHECK: Alignment should match open parenthesis'
