+++ src/include/sof/drivers/timer.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, Intel Corporation
+ * Copyright (c) 2018, Intel Corporation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,84 +25,32 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *
- * Author: Liam Girdwood <liam.r.girdwood@linux.intel.com>
+ * Author: Janusz Jankowski <janusz.jankowski@linux.intel.com>
  */
 
-#include <platform/timer.h>
-#include <platform/shim.h>
-#include <platform/interrupt.h>
-#include <sof/debug.h>
-#include <sof/audio/component.h>
-#include <stdint.h>
+#ifndef __INCLUDE_DRIVERS_TIMER__
+#define __INCLUDE_DRIVERS_TIMER__
 
-void platform_timer_start(struct timer *timer)
-{
-	arch_timer_enable(timer);
-}
+#include <sof/timer.h>
 
-void platform_timer_stop(struct timer *timer)
-{
-	arch_timer_disable(timer);
-}
+struct comp_dev;
+struct sof_ipc_stream_posn;
 
-int platform_timer_set(struct timer *timer, uint64_t ticks)
-{
-	return arch_timer_set(timer, ticks);
-}
-
-void platform_timer_clear(struct timer *timer)
-{
-	arch_timer_clear(timer);
-}
-
-uint64_t platform_timer_get(struct timer *timer)
-{
-	return arch_timer_get_system(timer);
-}
+int platform_timer_set(struct timer *timer, uint64_t ticks);
+void platform_timer_clear(struct timer *timer);
+uint64_t platform_timer_get(struct timer *timer);
+void platform_timer_start(struct timer *timer);
+void platform_timer_stop(struct timer *timer);
 
 /* get timestamp for host stream DMA position */
 void platform_host_timestamp(struct comp_dev *host,
-	struct sof_ipc_stream_posn *posn)
-{
-	int err;
-
-	/* get host postion */
-	err = comp_position(host, posn);
-	if (err == 0)
-		posn->flags |= SOF_TIME_HOST_VALID | SOF_TIME_HOST_64;
-}
+			     struct sof_ipc_stream_posn *posn);
 
 /* get timestamp for DAI stream DMA position */
 void platform_dai_timestamp(struct comp_dev *dai,
-	struct sof_ipc_stream_posn *posn)
-{
-	int err;
-
-	/* get DAI postion */
-	err = comp_position(dai, posn);
-	if (err == 0)
-		posn->flags |= SOF_TIME_DAI_VALID;
-
-	/* get SSP wallclock - DAI sets this to stream start value */
-	posn->wallclock = timer_get_system(platform_timer) - posn->wallclock;
-	posn->flags |= SOF_TIME_WALL_VALID | SOF_TIME_WALL_64;
-}
+			    struct sof_ipc_stream_posn *posn);
 
 /* get current wallclock for componnent */
-void platform_dai_wallclock(struct comp_dev *dai, uint64_t *wallclock)
-{
-	/* only 1 wallclock on HSW */
-	*wallclock = timer_get_system(platform_timer);
-}
+void platform_dai_wallclock(struct comp_dev *dai, uint64_t *wallclock);
 
-int timer_register(struct timer *timer, void(*handler)(void *arg), void *arg)
-{
-	switch (timer->id) {
-	case TIMER0:
-	case TIMER1:
-	case TIMER2:
-		return arch_timer_register(timer, handler, arg);
-	default:
-		return -EINVAL;
-	}
-}
+#endif
