+++ src/platform/baytrail/Makefile.am
@@ -8,7 +8,6 @@ libplatform_la_SOURCES = \
 	platform.c \
 	dai.c \
 	dma.c \
-	timer.c \
 	memory.c
 
 libplatform_la_CFLAGS = \
