+++ src/platform/haswell/include/platform/timer.h
@@ -43,24 +43,4 @@
 #define TIMER1		IRQ_NUM_TIMER2
 #define TIMER2		IRQ_NUM_TIMER3
 
-struct comp_dev;
-struct sof_ipc_stream_posn;
-
-int platform_timer_set(struct timer *timer, uint64_t ticks);
-void platform_timer_clear(struct timer *timer);
-uint64_t platform_timer_get(struct timer *timer);
-void platform_timer_start(struct timer *timer);
-void platform_timer_stop(struct timer *timer);
-
-/* get timestamp for host stream DMA position */
-void platform_host_timestamp(struct comp_dev *host,
-	struct sof_ipc_stream_posn *posn);
-
-/* get timestamp for DAI stream DMA position */
-void platform_dai_timestamp(struct comp_dev *dai,
-	struct sof_ipc_stream_posn *posn);
-
-/* get current wallclock for componnent */
-void platform_dai_wallclock(struct comp_dev *dai, uint64_t *wallclock);
-
 #endif
